Add files via upload
In this update and modification, I fixed the order in which the previous posts were added and displayed them on the home page in chronological order. And added that all posts contain location information generated based on GPS (in the app test preview, I set it to generate posts that simulate other users sending to the server) and virtual user names, and retain the classification and labelling functions of posts in the system (can be changed and added). You can root Add according to demand. This 2.0 update mainly ensures that there is a stable version at the time of Presentation to introduce the final effect of our software.